feat: Add user profile editor modal.
This also includes some things that should've been in their own commit, but it's pretty late and I want to go to bed. Doesn't hurt anyhow, but I prefer smaller commits.

This also includes updated database types, an updated PREREQUISITES to match the namechange for the SQL profile functions, and refactored the SQL profile functions in-code to match their new name.